MB-56973 [5/6]: Increase frequency of ConnManager from 1Hz to 10Hz
To be able to send DCP noop messages every 100ms, we need the task
which checks if they should be sent - ConnManager - to be scheduled
every 100ms. Change the 'connection_manager_interval' config param
from integer to float, and change the default to 0.1s

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/195728
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/ep_engine.cc
@@ -919,7 +919,7 @@ cb::engine_errc EventuallyPersistentEngine::setDcpParam(const std::string& key,
         } else if (key == "dcp_consumer_buffer_ratio") {
             getConfiguration().setDcpConsumerBufferRatio(std::stof(val));
         } else if (key == "connection_manager_interval") {
-            getConfiguration().setConnectionManagerInterval(std::stoull(val));
+            getConfiguration().setConnectionManagerInterval(std::stof(val));
         } else if (key ==
                    "dcp_consumer_process_buffered_messages_yield_limit") {
             getConfiguration().setDcpConsumerProcessBufferedMessagesYieldLimit(
